backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.26.4
openai==0.27.0
packaging==24.2
preshed==3.0.9
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.2
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
setuptools==75.8.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.3
thinc==8.3.4
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
